+++ src/containers/ModuleLists.js
@@ -15,6 +15,7 @@ export default class ModuleLists extends React.Component {
         this.setModuleTitle = this.setModuleTitle.bind(this);
         this.createModule = this.createModule.bind(this);
         this.deleteModule = this.deleteModule.bind(this);
+        this.createModuleServiceCall = this.createModuleServiceCall.bind(this);
     }
 
     setCourseId(courseId) {
@@ -22,17 +23,17 @@ export default class ModuleLists extends React.Component {
     }
 
     setModuleTitle(event) {
-        this.setState({module: {title: event.target.value}})
+        this.setState({module: {title: event.target.value}});
     }
 
     componentDidMount() {
         this.setCourseId(this.props.courseId);
-        this.findAllModulesForCourse(this.props.courseId)
+        this.findAllModulesForCourse(this.props.courseId);
     }
 
     componentWillReceiveProps(newProps) {
         this.setCourseId(newProps.courseId);
-        this.findAllModulesForCourse(newProps.courseId)
+        this.findAllModulesForCourse(newProps.courseId);
     }
 
     findAllModulesForCourse(courseId) {
@@ -42,19 +43,32 @@ export default class ModuleLists extends React.Component {
     }
 
     setModules(modules) {
+        this.setState({module: {title: ''}});
         this.setState({modules: modules})
     }
 
     createModule() {
+        if (undefined === this.state.module || '' === this.state.module.title) {
+            this.setState({module: {title: 'New Module'}}, function () {
+                this.createModuleServiceCall();
+            });
+        }
+        else this.createModuleServiceCall();
+    }
+
+    createModuleServiceCall() {
         this.moduleService.createModule(this.state.courseId, this.state.module).then(() => {
             this.findAllModulesForCourse(this.state.courseId);
         });
     }
 
     deleteModule(moduleId) {
-        this.moduleService.deleteModule(moduleId).then(() => {
-            this.findAllModulesForCourse(this.state.courseId)
-        });
+        var input = window.confirm("Are you sure you want to delete this module?");
+        if (input === true) {
+            this.moduleService.deleteModule(moduleId).then(() => {
+                this.findAllModulesForCourse(this.state.courseId)
+            });
+        }
     }
 
     renderModules() {
@@ -65,28 +79,28 @@ export default class ModuleLists extends React.Component {
                                    delete={this.deleteModule}
                                    update={this.updateModule}/>
         });
-        return (<ul>{modules}</ul>)
+        return modules;
     }
 
     render() {
         return (
-            <div className="container-fluid ">
-                <div>
-                    <h4>Modules for courseId: {this.state.courseId}</h4>
-                </div>
-                <div className="row">
-                    <input className="form-control" value={this.state.module.title}
-                           placeholder="New Module"
-                           onChange={this.setModuleTitle}/>
-                    <span className="input-group-addon">
-                        <i className="fa fa-plus" onClick={this.createModule}/>
-                    </span>
-                </div>
-                <div>
-                    <ul className="list-group">
-                        {this.renderModules()}
-                    </ul>
-                </div>
+            <div className="wbdv-module-list">
+                <h2>&nbsp;&nbsp;&nbsp;&nbsp;Modules for {this.state.courseId}</h2>
+                <ul>
+                    <div className="row">
+                        <div className="col-11">
+                            <input onChange={this.setModuleTitle}
+                                   value={this.state.module.title}
+                                   placeholder="New Module"
+                                   className="form-control"/>
+                        </div>
+                        <div className="col-1">
+                            <i className="fa fa-plus" onClick={this.createModule}/>
+                        </div>
+                    </div>
+
+                    {this.renderModules()}
+                </ul>
             </div>
         );
     }
